+++ src/main/java/com/christopherosthues/starwarsprogressbar/ui/StarWarsProgressBarFactory.java
@@ -1,5 +1,6 @@
 package com.christopherosthues.starwarsprogressbar.ui;
 
+import com.christopherosthues.starwarsprogressbar.models.StarWarsVehicle;
 import com.intellij.util.ui.JBUI;
 
 import javax.swing.*;
@@ -10,6 +11,6 @@ public class StarWarsProgressBarFactory extends BasicProgressBarUI {
     @SuppressWarnings( {"MethodOverridesStaticMethodOfSuperclass", "UnusedDeclaration"})
     public static ComponentUI createUI(final JComponent c) {
         c.setBorder(JBUI.Borders.empty().asUIResource());
-        return new StarWarsProgressBarUI();
+        return new StarWarsProgressBarUI(() -> null, new StarWarsVehicle("", "", 0, 0, 0));
     }
 }