+++ src/main/kotlin/com/christopherosthues/starwarsprogressbar/ui/LightsaberColor.kt
@@ -5,8 +5,28 @@ import java.awt.Color
 
 internal object LightsaberColor {
     val Cyan : JBColor = JBColor(Color(0, 255, 255), Color(0, 255, 255))
+    val LightBlue : JBColor = JBColor(Color(0, 255, 255), Color(0, 255, 255))
+    val Blue : JBColor = JBColor(Color(0, 255, 255), Color(0, 255, 255))
+    val DarkBlue : JBColor = JBColor(Color(0, 255, 255), Color(0, 255, 255))
+    val LightGreen : JBColor = JBColor(Color(0, 255, 255), Color(0, 255, 255))
+    val Green : JBColor = JBColor(Color(0, 255, 255), Color(0, 255, 255))
+    val DarkGreen : JBColor = JBColor(Color(0, 255, 255), Color(0, 255, 255))
+    val Orange : JBColor = JBColor(Color(0, 255, 255), Color(0, 255, 255))
+    val Yellow : JBColor = JBColor(Color(0, 255, 255), Color(0, 255, 255))
+    val Purple : JBColor = JBColor(Color(0, 255, 255), Color(0, 255, 255))
+    val White : JBColor = JBColor(Color(0, 255, 255), Color(0, 255, 255))
 
     val colors = mapOf(
-        "cyan" to Cyan
+        "cyan" to Cyan,
+        "lightBlue" to LightBlue,
+        "blue" to Blue,
+        "darkBlue" to DarkBlue,
+        "lightGreen" to LightGreen,
+        "green" to Green,
+        "darkGreen" to DarkGreen,
+        "orange" to Orange,
+        "yellow" to Yellow,
+        "purple" to Purple,
+        "white" to White,
     )
 }